[#349] Add instantiated node custom service
Problem: We want to be able to start services that will run nodes using
different node configs. Existing tezos-node-custom isn't flexible enough
to be used in the baking service.

Solution: Add instanted node custom service in which service
dependencies and the location of the environment file will be determined
by the instance name. Instances of such service can be used as a part of
custom baking services.