[#349] Add instantiated custom baking service
Problem: We want to be able to run baking services on custom networks.

Solution: Add instantiated custom baking service. Its instances can be
used to run baking services on a network with arbitrary node config.

This service requires all existing daemon services, so instances of it
can be safely run on a network with an arbitrary existing protocol
supported by our packages.