[#237] Refactor common wizard part
Problem: we now have two wizard scripts, but there's a lot of copy-paste
in them. It's less maintainable and nice.

Solution: moved the common parts, such as validators and Step and Setup
class structures to a separate file that the wizards use. Also improved
readability in the Setup Wizard some.