[#522] Switch node service to 'Type=notify'
Problem: Currently node services is considered to be up even though it
is yet unable to respond to RPC queries. This requires additional
workarounds for services that depend on node.

Solution: Switch node service to 'Type=notify' and mark service as ready
only once it start to respond to RPC queries.