[#567] Add a separate doc for Windows setups
Problem: using WSL it's possible to have a working setup on
Windows systems. We document this, but this documentation was add
over time and is split in different documents.
This is confusing and makes it hard for users to find all the
necessary info and steps to follow.

Solution: collect all the documentation for Windows in a single
document. Additionally reword some instructions.